refactor(qc): delete three dead legacy routes (Tier D)
Verified dead three ways before removal: no frontend caller (UI uses the HTMX
paths), no app-code caller, and zero hits in 14 days of live access logs.

- PUT /api/quotes/{id} + POST /api/quotes/{id}/revise (crm/quotes.py) — the
  legacy JSON editor/revise pair that diverged from the HTMX path (QC M2).
  QuoteUpdate schema removed with them. The JSON reopen route survives (its
  own scope test now pins the shared get_quote_for_user gate via /result,
  and the authz allow-side pin moved to the DELETE route).
- POST /v2/prepayments (prepayments.py) — the JSON create that never sent the
  accounting notice (QC #11); the HTMX create is the only live path and now
  carries the G5 ownership authz pins.
- excess_service.import_line_items — the unguarded legacy twin of
  preview_import/confirm_import, unreachable by any route (QC #33); moved to
  tests/fixtures/excess_import.py for the ~15 tests that seed with it.

Tests updated: dead-route tests removed; scope/authz pins repointed to
surviving routes. Full suite: 24,317 passed (the 1 failure is the pre-existing
revenue_90d time bomb, fixed separately in #886).
