+++ tests/test_company_merge_service.py
@@ -10,7 +10,8 @@ from app.models import Company, CustomerSite, User
 from app.models.sourcing import Requisition, Sighting
 from app.services.company_merge_service import delete_companies, merge_companies
 from app.services.excess_mirror import _virtual_req_name, publish_list
-from app.services.excess_service import create_excess_list, import_line_items
+from app.services.excess_service import create_excess_list
+from tests.fixtures.excess_import import import_line_items
 
 
 def _make_pair(db_session, keep_kwargs, remove_kwargs):
