+++ tests/test_resell_draft_edit.py
@@ -27,7 +27,8 @@ from app.models import Company, User
 from app.models.excess import ExcessLineItem, ExcessList
 from app.models.intelligence import MaterialCard
 from app.services import excess_service
-from app.services.excess_service import create_excess_list, import_line_items
+from app.services.excess_service import create_excess_list
+from tests.fixtures.excess_import import import_line_items
 
 # ── Fixtures ─────────────────────────────────────────────────────────
